+++ src/chemex/models/kinetic/settings_2st_rs.py
@@ -13,6 +13,7 @@ def make_settings_2st_rs(_conditions: Conditions) -> dict[str, ParamLocalSetting
             name_setting=NameSetting("kex_ab", "g", TPL),
             value=200.0,
             min=0.0,
+            max=1.0e6,
             vary=True,
         ),
         "pb": ParamLocalSetting(